+++ model/tokenGrant.go
@@ -4,7 +4,7 @@ import "gorm.io/gorm"
 
 type TokenGrant struct {
 	gorm.Model
-	AccountId   uint64 		`gorm:"index"`
+	AccountId   uint 		`gorm:"index"`
 	Date		string		`gorm:"index"`
 	Token		uint64
 	Way			int
